Top-level Reader
This isn't ideal, but probably the closest we can get to my original
goal. As paf31 describes here, the inner Monad is to allow event
handlers to run effectful code, the render function, however, has no
access to it:

https://stackoverflow.com/questions/30649310/monad-transformer-for-halogen-components

Since SF1 is not a Monad at the moment, we can't easily wrap it or
extend it through a MonadReader. Maybe I can revisit this at some point
in the future.